Catalan

Documentation

data Nat

Natural numbers...

Constructors

Z

S Nat

type family Plus m n :: Nat

Type family allowing us to add natural numbers at the type level

data NNPath height where

NNPath (non-negative path) are paths consisting of up (U) and down (D) segments, which start at zero and never go below zero

Constructors

End :: NNPath Z

U :: NNPath m -> NNPath (S m)

D :: NNPath (S m) -> NNPath m

(-) :: NNPath x -> (NNPath x -> NNPath y) -> NNPath y

This operator lets you construct paths in a left-to-right manner, e.g. End-U-D-U-U-D-D

type Dyck = NNPath Z

A Dyck path is a non-negative path ending at zero

(|+|) :: NNPath n -> NNPath m -> NNPath (Plus m n)

Non-negative paths may be concatenated, resulting in another non-negative path

class CutZ x where

Class to allow slicing paths at the zero crossing; used in split below

Methods

cut :: NNPath x -> (NNPath x, Dyck)

split :: Dyck -> [Dyck]

A Dyck path may be separated into a list of Dyck paths. This operation guarantees that each Dyck path in the resulting list will be a *prime* Dyck path (i.e. no zero-crossings other than start and end), but we don't yet have a way of meaningfully encoding this in the type system so that we can take advantage of it to type-safely implement an unbump operation for prime Dyck paths.

The following identities hold:

 foldr1 (|+|) . split === id

 split . foldr1 (|+|) === id

tack :: x -> [x] -> [x]

Tack an element onto the end of a list

bump :: Dyck -> Dyck

Take a Dyck path and return it with an up before it and a down after it, thus bumping it upwards

class Catalan a where

Typeclass for things countable by Catalan numbers

 toDyck . fromDyck === id

 fromDyck . toDyck === id

Methods

toDyck :: a -> Dyck

fromDyck :: Dyck -> a

data Tree

Trees with any number of child nodes can be counted by Catalan numbers

Constructors

Node [Tree]

children :: Tree -> [Tree]

Returns the list of children for a tree

data Dir

A direction is either up (Up) or down (Dn)

Constructors

Up

Dn

plusDir :: Enum a => Dir -> a -> a

Adds either 1 or -1 based on the direction given

isDyckList :: [Dir] -> Bool

Determines if a list of directions constitutes a Dyck path

parseDyck :: [Dir] -> Maybe Dyck

We can take a purely value-level, no-nonsense list of ordinary directions and (maybe) parse it into a Dyck path -- if and only if it represents a Dyck path. This uses the same tree-zipper technique used in implementing the fromDyck method of trees

Note that (for Dyck paths ONLY -- other inputs yield bottom):

 unparseNNPath . fromJust . parseDyck === id

 fromJust . parseDyck . unparseNNPath === id

unparseNNPath :: NNPath x -> [Dir]

Takes a non-negative path and unparses it, returning just a sequence of directions